# perf: specialize match function at UPat init [run_process_replay]