⚠️ This is probably too complex for the 50% speed increase it offers on the `_match` function, but I wanted to share the idea. 

At `UPat` initialization, the `_match` function can be specialized based on which elements of the `UPat` are not None.
This cuts down on unnecessary checks and attribute access.

```
PROFILE=0 python -O test/external/external_benchmark_schedule.py
master
***** model forward in  15.68 ms
***** model schedule in   5.22 ms
***** model lower in 575.02 ms
n:  331816  tm: 210.41ms  tot: 371.24ms uops.py:111:_match                                 <-  66% uops.py:141:rewrite

branch
***** model forward in  15.61 ms
***** model schedule in   5.02 ms
***** model lower in 525.86 ms
n:  331816  tm:  72.74ms  tot: 254.60ms uops.py:134:_match                                 <-  57% uops.py:122:src_unmatch
```

Summary of changes:
  * reduce `_match` to a tuple of `UPat`-specific match functions that only includes the relevant functions. This way, all member lookups are also removed from the matching calls.
  * move permutations from the `src` match (each match call) to the `UPat` init (once)
  * early stopping for the src children match
  * `op`, `dtype` and `arg` matches are simplified to `in tuple`

I tested real functions instead of lambda functions but that does not seem to matter.


Example:

`UPat(UOps.CONST, name="mval")` has `unmatch_funcs=(lambda uop,_: uop.op not in(UOps.CONST,), lambda uop,store: store.setdefault("mval", uop) is not uop)`