@DKormann 2024-07-02
why is it faster? because the lambdas or because you have the different loop for src_unmatch