@DKormann 2024-07-02
oh i see the any( ... ) goes through all children that makes sense