@DKormann 2024-07-02
this seems like you could get it passed easy when split into independent prs no? also i dont see a reason why this cant be done in almost no additional lines