@DKormann 2024-07-02
well than try to pr at least the idea with lambda functions