@geohot 2024-07-02
How are you getting 575 ms baseline? On an M3 Pro

```
diane@blak:~/tinygrad$ PROFILE=0 python -O test/external/external_benchmark_schedule.py
***** model forward in  18.74 ms
***** model schedule in   6.11 ms
***** model lower in 872.02 ms
```

I'm okay with the tuple refactor, I just think it should be like 3 lines and not make the syntax worse.