@roelofvandijk 2024-07-02
Python 3.12, M3 Pro