@roelofvandijk 2024-07-09
Closing this, might pull out only the all-tuple init